+++ src/Presentation/Nop.Web/Factories/CommonModelFactory.cs
@@ -1,0 +1,687 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Web;
+using System.Web.Mvc;
+using Nop.Core;
+using Nop.Core.Caching;
+using Nop.Core.Domain;
+using Nop.Core.Domain.Blogs;
+using Nop.Core.Domain.Catalog;
+using Nop.Core.Domain.Common;
+using Nop.Core.Domain.Customers;
+using Nop.Core.Domain.Forums;
+using Nop.Core.Domain.Localization;
+using Nop.Core.Domain.News;
+using Nop.Core.Domain.Orders;
+using Nop.Core.Domain.Vendors;
+using Nop.Services.Catalog;
+using Nop.Services.Common;
+using Nop.Services.Customers;
+using Nop.Services.Directory;
+using Nop.Services.Forums;
+using Nop.Services.Localization;
+using Nop.Services.Media;
+using Nop.Services.Orders;
+using Nop.Services.Security;
+using Nop.Services.Seo;
+using Nop.Services.Topics;
+using Nop.Web.Extensions;
+using Nop.Web.Framework.Security.Captcha;
+using Nop.Web.Framework.Themes;
+using Nop.Web.Framework.UI;
+using Nop.Web.Infrastructure.Cache;
+using Nop.Web.Models.Catalog;
+using Nop.Web.Models.Common;
+using Nop.Web.Models.Topics;
+
+namespace Nop.Web.Factories
+{
+    public partial class CommonModelFactory : ICommonModelFactory
+    {
+        #region Fields
+
+        private readonly ICategoryService _categoryService;
+        private readonly IProductService _productService;
+        private readonly IManufacturerService _manufacturerService;
+        private readonly ITopicService _topicService;
+        private readonly ILanguageService _languageService;
+        private readonly ICurrencyService _currencyService;
+        private readonly ILocalizationService _localizationService;
+        private readonly IWorkContext _workContext;
+        private readonly IStoreContext _storeContext;
+        private readonly ISitemapGenerator _sitemapGenerator;
+        private readonly IThemeContext _themeContext;
+        private readonly IThemeProvider _themeProvider;
+        private readonly IForumService _forumservice;
+        private readonly IGenericAttributeService _genericAttributeService;
+        private readonly IWebHelper _webHelper;
+        private readonly IPermissionService _permissionService;
+        private readonly ICacheManager _cacheManager;
+        private readonly IPageHeadBuilder _pageHeadBuilder;
+        private readonly IPictureService _pictureService;
+        private readonly HttpContextBase _httpContext;
+
+        private readonly CatalogSettings _catalogSettings;
+        private readonly StoreInformationSettings _storeInformationSettings;
+        private readonly CommonSettings _commonSettings;
+        private readonly BlogSettings _blogSettings;
+        private readonly NewsSettings _newsSettings;
+        private readonly ForumSettings _forumSettings;
+        private readonly LocalizationSettings _localizationSettings;
+        private readonly CaptchaSettings _captchaSettings;
+        private readonly VendorSettings _vendorSettings;
+
+        #endregion
+
+        #region Constructors
+
+        public CommonModelFactory(ICategoryService categoryService,
+            IProductService productService,
+            IManufacturerService manufacturerService,
+            ITopicService topicService,
+            ILanguageService languageService,
+            ICurrencyService currencyService,
+            ILocalizationService localizationService,
+            IWorkContext workContext,
+            IStoreContext storeContext,
+            ISitemapGenerator sitemapGenerator,
+            IThemeContext themeContext,
+            IThemeProvider themeProvider,
+            IForumService forumService,
+            IGenericAttributeService genericAttributeService,
+            IWebHelper webHelper,
+            IPermissionService permissionService,
+            ICacheManager cacheManager,
+            IPageHeadBuilder pageHeadBuilder,
+            IPictureService pictureService,
+            HttpContextBase httpContext,
+            CatalogSettings catalogSettings,
+            StoreInformationSettings storeInformationSettings,
+            CommonSettings commonSettings,
+            BlogSettings blogSettings,
+            NewsSettings newsSettings,
+            ForumSettings forumSettings,
+            LocalizationSettings localizationSettings,
+            CaptchaSettings captchaSettings,
+            VendorSettings vendorSettings)
+        {
+            this._categoryService = categoryService;
+            this._productService = productService;
+            this._manufacturerService = manufacturerService;
+            this._topicService = topicService;
+            this._languageService = languageService;
+            this._currencyService = currencyService;
+            this._localizationService = localizationService;
+            this._workContext = workContext;
+            this._storeContext = storeContext;
+            this._sitemapGenerator = sitemapGenerator;
+            this._themeContext = themeContext;
+            this._themeProvider = themeProvider;
+            this._forumservice = forumService;
+            this._genericAttributeService = genericAttributeService;
+            this._webHelper = webHelper;
+            this._permissionService = permissionService;
+            this._cacheManager = cacheManager;
+            this._pageHeadBuilder = pageHeadBuilder;
+            this._pictureService = pictureService;
+            this._httpContext = httpContext;
+
+            this._catalogSettings = catalogSettings;
+            this._storeInformationSettings = storeInformationSettings;
+            this._commonSettings = commonSettings;
+            this._blogSettings = blogSettings;
+            this._newsSettings = newsSettings;
+            this._forumSettings = forumSettings;
+            this._localizationSettings = localizationSettings;
+            this._captchaSettings = captchaSettings;
+            this._vendorSettings = vendorSettings;
+        }
+
+        #endregion
+
+        #region Utilities
+
+        protected virtual int GetUnreadPrivateMessages()
+        {
+            var result = 0;
+            var customer = _workContext.CurrentCustomer;
+            if (_forumSettings.AllowPrivateMessages && !customer.IsGuest())
+            {
+                var privateMessages = _forumservice.GetAllPrivateMessages(_storeContext.CurrentStore.Id,
+                    0, customer.Id, false, null, false, string.Empty, 0, 1);
+
+                if (privateMessages.TotalCount > 0)
+                {
+                    result = privateMessages.TotalCount;
+                }
+            }
+
+            return result;
+        }
+
+        #endregion
+
+        #region Methods
+
+        public virtual LogoModel PrepareLogoModel()
+        {
+            var model = new LogoModel
+            {
+                StoreName = _storeContext.CurrentStore.GetLocalized(x => x.Name)
+            };
+
+            var cacheKey = string.Format(ModelCacheEventConsumer.STORE_LOGO_PATH, _storeContext.CurrentStore.Id, _themeContext.WorkingThemeName, _webHelper.IsCurrentConnectionSecured());
+            model.LogoPath = _cacheManager.Get(cacheKey, () =>
+            {
+                var logo = "";
+                var logoPictureId = _storeInformationSettings.LogoPictureId;
+                if (logoPictureId > 0)
+                {
+                    logo = _pictureService.GetPictureUrl(logoPictureId, showDefaultPicture: false);
+                }
+                if (String.IsNullOrEmpty(logo))
+                {
+                    //use default logo
+                    logo = string.Format("{0}Themes/{1}/Content/images/logo.png", _webHelper.GetStoreLocation(), _themeContext.WorkingThemeName);
+                }
+                return logo;
+            });
+
+            return model;
+        }
+
+        public virtual LanguageSelectorModel PrepareLanguageSelectorModel()
+        {
+            var availableLanguages = _cacheManager.Get(string.Format(ModelCacheEventConsumer.AVAILABLE_LANGUAGES_MODEL_KEY, _storeContext.CurrentStore.Id), () =>
+            {
+                var result = _languageService
+                    .GetAllLanguages(storeId: _storeContext.CurrentStore.Id)
+                    .Select(x => new LanguageModel
+                    {
+                        Id = x.Id,
+                        Name = x.Name,
+                        FlagImageFileName = x.FlagImageFileName,
+                    })
+                    .ToList();
+                return result;
+            });
+
+            var model = new LanguageSelectorModel
+            {
+                CurrentLanguageId = _workContext.WorkingLanguage.Id,
+                AvailableLanguages = availableLanguages,
+                UseImages = _localizationSettings.UseImagesForLanguageSelection
+            };
+            
+            return model;
+        }
+
+        public virtual CurrencySelectorModel PrepareCurrencySelectorModel()
+        {
+            var availableCurrencies = _cacheManager.Get(string.Format(ModelCacheEventConsumer.AVAILABLE_CURRENCIES_MODEL_KEY, _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id), () =>
+            {
+                var result = _currencyService
+                    .GetAllCurrencies(storeId: _storeContext.CurrentStore.Id)
+                    .Select(x =>
+                    {
+                        //currency char
+                        var currencySymbol = "";
+                        if (!string.IsNullOrEmpty(x.DisplayLocale))
+                            currencySymbol = new RegionInfo(x.DisplayLocale).CurrencySymbol;
+                        else
+                            currencySymbol = x.CurrencyCode;
+                        //model
+                        var currencyModel = new CurrencyModel
+                        {
+                            Id = x.Id,
+                            Name = x.GetLocalized(y => y.Name),
+                            CurrencySymbol = currencySymbol
+                        };
+                        return currencyModel;
+                    })
+                    .ToList();
+                return result;
+            });
+
+            var model = new CurrencySelectorModel
+            {
+                CurrentCurrencyId = _workContext.WorkingCurrency.Id,
+                AvailableCurrencies = availableCurrencies
+            };
+
+            return model;
+        }
+
+        public virtual TaxTypeSelectorModel PrepareTaxTypeSelectorModel()
+        {
+            var model = new TaxTypeSelectorModel
+            {
+                CurrentTaxType = _workContext.TaxDisplayType
+            };
+
+            return model;
+        }
+
+        public virtual HeaderLinksModel PrepareHeaderLinksModel()
+        {
+            var customer = _workContext.CurrentCustomer;
+
+            var unreadMessageCount = GetUnreadPrivateMessages();
+            var unreadMessage = string.Empty;
+            var alertMessage = string.Empty;
+            if (unreadMessageCount > 0)
+            {
+                unreadMessage = string.Format(_localizationService.GetResource("PrivateMessages.TotalUnread"), unreadMessageCount);
+
+                //notifications here
+                if (_forumSettings.ShowAlertForPM &&
+                    !customer.GetAttribute<bool>(SystemCustomerAttributeNames.NotifiedAboutNewPrivateMessages, _storeContext.CurrentStore.Id))
+                {
+                    _genericAttributeService.SaveAttribute(customer, SystemCustomerAttributeNames.NotifiedAboutNewPrivateMessages, true, _storeContext.CurrentStore.Id);
+                    alertMessage = string.Format(_localizationService.GetResource("PrivateMessages.YouHaveUnreadPM"), unreadMessageCount);
+                }
+            }
+
+            var model = new HeaderLinksModel
+            {
+                IsAuthenticated = customer.IsRegistered(),
+                CustomerName = customer.IsRegistered() ? customer.FormatUserName() : "",
+                ShoppingCartEnabled = _permissionService.Authorize(StandardPermissionProvider.EnableShoppingCart),
+                WishlistEnabled = _permissionService.Authorize(StandardPermissionProvider.EnableWishlist),
+                AllowPrivateMessages = customer.IsRegistered() && _forumSettings.AllowPrivateMessages,
+                UnreadPrivateMessages = unreadMessage,
+                AlertMessage = alertMessage,
+            };
+            //performance optimization (use "HasShoppingCartItems" property)
+            if (customer.HasShoppingCartItems)
+            {
+                model.ShoppingCartItems = customer.ShoppingCartItems
+                    .Where(sci => sci.ShoppingCartType == ShoppingCartType.ShoppingCart)
+                    .LimitPerStore(_storeContext.CurrentStore.Id)
+                    .ToList()
+                    .GetTotalProducts();
+                model.WishlistItems = customer.ShoppingCartItems
+                    .Where(sci => sci.ShoppingCartType == ShoppingCartType.Wishlist)
+                    .LimitPerStore(_storeContext.CurrentStore.Id)
+                    .ToList()
+                    .GetTotalProducts();
+            }
+
+            return model;
+        }
+
+        public virtual AdminHeaderLinksModel PrepareAdminHeaderLinksModel()
+        {
+            var customer = _workContext.CurrentCustomer;
+
+            var model = new AdminHeaderLinksModel
+            {
+                ImpersonatedCustomerName = customer.IsRegistered() ? customer.FormatUserName() : "",
+                IsCustomerImpersonated = _workContext.OriginalCustomerIfImpersonated != null,
+                DisplayAdminLink = _permissionService.Authorize(StandardPermissionProvider.AccessAdminPanel),
+                EditPageUrl = _pageHeadBuilder.GetEditPageUrl()
+            };
+
+            return model;
+        }
+
+        public virtual SocialModel PrepareSocialModel()
+        {
+            var model = new SocialModel
+            {
+                FacebookLink = _storeInformationSettings.FacebookLink,
+                TwitterLink = _storeInformationSettings.TwitterLink,
+                YoutubeLink = _storeInformationSettings.YoutubeLink,
+                GooglePlusLink = _storeInformationSettings.GooglePlusLink,
+                WorkingLanguageId = _workContext.WorkingLanguage.Id,
+                NewsEnabled = _newsSettings.Enabled,
+            };
+
+            return model;
+        }
+
+        public virtual FooterModel PrepareFooterModel()
+        {
+            //footer topics
+            string topicCacheKey = string.Format(ModelCacheEventConsumer.TOPIC_FOOTER_MODEL_KEY,
+                _workContext.WorkingLanguage.Id,
+                _storeContext.CurrentStore.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()));
+            var cachedTopicModel = _cacheManager.Get(topicCacheKey, () =>
+                _topicService.GetAllTopics(_storeContext.CurrentStore.Id)
+                .Where(t => t.IncludeInFooterColumn1 || t.IncludeInFooterColumn2 || t.IncludeInFooterColumn3)
+                .Select(t => new FooterModel.FooterTopicModel
+                {
+                    Id = t.Id,
+                    Name = t.GetLocalized(x => x.Title),
+                    SeName = t.GetSeName(),
+                    IncludeInFooterColumn1 = t.IncludeInFooterColumn1,
+                    IncludeInFooterColumn2 = t.IncludeInFooterColumn2,
+                    IncludeInFooterColumn3 = t.IncludeInFooterColumn3
+                })
+                .ToList()
+            );
+
+            //model
+            var model = new FooterModel
+            {
+                StoreName = _storeContext.CurrentStore.GetLocalized(x => x.Name),
+                WishlistEnabled = _permissionService.Authorize(StandardPermissionProvider.EnableWishlist),
+                ShoppingCartEnabled = _permissionService.Authorize(StandardPermissionProvider.EnableShoppingCart),
+                SitemapEnabled = _commonSettings.SitemapEnabled,
+                WorkingLanguageId = _workContext.WorkingLanguage.Id,
+                BlogEnabled = _blogSettings.Enabled,
+                CompareProductsEnabled = _catalogSettings.CompareProductsEnabled,
+                ForumEnabled = _forumSettings.ForumsEnabled,
+                NewsEnabled = _newsSettings.Enabled,
+                RecentlyViewedProductsEnabled = _catalogSettings.RecentlyViewedProductsEnabled,
+                NewProductsEnabled = _catalogSettings.NewProductsEnabled,
+                DisplayTaxShippingInfoFooter = _catalogSettings.DisplayTaxShippingInfoFooter,
+                HidePoweredByNopCommerce = _storeInformationSettings.HidePoweredByNopCommerce,
+                AllowCustomersToApplyForVendorAccount = _vendorSettings.AllowCustomersToApplyForVendorAccount,
+                Topics = cachedTopicModel
+            };
+
+            return model;
+        }
+        
+        public virtual ContactUsModel PrepareContactUsModel(ContactUsModel model, bool excludeProperties)
+        {
+            if (model == null)
+                throw new ArgumentNullException("model");
+
+            if (!excludeProperties)
+            {
+                model.Email = _workContext.CurrentCustomer.Email;
+                model.FullName = _workContext.CurrentCustomer.GetFullName();
+            }
+            model.SubjectEnabled = _commonSettings.SubjectFieldOnContactUsForm;
+            model.DisplayCaptcha = _captchaSettings.Enabled && _captchaSettings.ShowOnContactUsPage;
+            
+            return model;
+        }
+
+        public virtual ContactVendorModel PrepareContactVendorModel(ContactVendorModel model, Vendor vendor, bool excludeProperties)
+        {
+            if (model == null)
+                throw new ArgumentNullException("model");
+
+            if (vendor == null)
+                throw new ArgumentNullException("vendor");
+
+            if (!excludeProperties)
+            {
+                model.Email = _workContext.CurrentCustomer.Email;
+                model.FullName = _workContext.CurrentCustomer.GetFullName();
+            }
+
+            model.SubjectEnabled = _commonSettings.SubjectFieldOnContactUsForm;
+            model.DisplayCaptcha = _captchaSettings.Enabled && _captchaSettings.ShowOnContactUsPage;
+            model.VendorId = vendor.Id;
+            model.VendorName = vendor.GetLocalized(x => x.Name);
+            
+            return model;
+        }
+
+        public virtual SitemapModel PrepareSitemapModel()
+        {
+            string cacheKey = string.Format(ModelCacheEventConsumer.SITEMAP_PAGE_MODEL_KEY,
+                _workContext.WorkingLanguage.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                _storeContext.CurrentStore.Id);
+            var cachedModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var model = new SitemapModel
+                {
+                    BlogEnabled = _blogSettings.Enabled,
+                    ForumEnabled = _forumSettings.ForumsEnabled,
+                    NewsEnabled = _newsSettings.Enabled,
+                };
+                //categories
+                if (_commonSettings.SitemapIncludeCategories)
+                {
+                    var categories = _categoryService.GetAllCategories(storeId: _storeContext.CurrentStore.Id);
+                    model.Categories = categories.Select(x => x.ToModel()).ToList();
+                }
+                //manufacturers
+                if (_commonSettings.SitemapIncludeManufacturers)
+                {
+                    var manufacturers = _manufacturerService.GetAllManufacturers(storeId: _storeContext.CurrentStore.Id);
+                    model.Manufacturers = manufacturers.Select(x => x.ToModel()).ToList();
+                }
+                //products
+                if (_commonSettings.SitemapIncludeProducts)
+                {
+                    //limit product to 200 until paging is supported on this page
+                    var products = _productService.SearchProducts(storeId: _storeContext.CurrentStore.Id,
+                        visibleIndividuallyOnly: true,
+                        pageSize: 200);
+                    model.Products = products.Select(product => new ProductOverviewModel
+                    {
+                        Id = product.Id,
+                        Name = product.GetLocalized(x => x.Name),
+                        ShortDescription = product.GetLocalized(x => x.ShortDescription),
+                        FullDescription = product.GetLocalized(x => x.FullDescription),
+                        SeName = product.GetSeName(),
+                    }).ToList();
+                }
+
+                //topics
+                var topics = _topicService.GetAllTopics(_storeContext.CurrentStore.Id)
+                    .Where(t => t.IncludeInSitemap)
+                    .ToList();
+                model.Topics = topics.Select(topic => new TopicModel
+                {
+                    Id = topic.Id,
+                    SystemName = topic.SystemName,
+                    IncludeInSitemap = topic.IncludeInSitemap,
+                    IsPasswordProtected = topic.IsPasswordProtected,
+                    Title = topic.GetLocalized(x => x.Title),
+                })
+                .ToList();
+                return model;
+            });
+
+            return cachedModel;
+        }
+
+        public virtual string PrepareSitemapXml(UrlHelper url)
+        {
+            string cacheKey = string.Format(ModelCacheEventConsumer.SITEMAP_SEO_MODEL_KEY,
+                _workContext.WorkingLanguage.Id,
+                string.Join(",", _workContext.CurrentCustomer.GetCustomerRoleIds()),
+                _storeContext.CurrentStore.Id);
+            var siteMap = _cacheManager.Get(cacheKey, () => _sitemapGenerator.Generate(url));
+            return siteMap;
+        }
+        
+        public virtual StoreThemeSelectorModel PrepareStoreThemeSelectorModel()
+        {
+            var model = new StoreThemeSelectorModel();
+            var currentTheme = _themeProvider.GetThemeConfiguration(_themeContext.WorkingThemeName);
+            model.CurrentStoreTheme = new StoreThemeModel
+            {
+                Name = currentTheme.ThemeName,
+                Title = currentTheme.ThemeTitle
+            };
+            model.AvailableStoreThemes = _themeProvider.GetThemeConfigurations()
+                .Select(x => new StoreThemeModel
+                {
+                    Name = x.ThemeName,
+                    Title = x.ThemeTitle
+                })
+                .ToList();
+            return model;
+        }
+        
+        public virtual FaviconModel PrepareFaviconModel()
+        {
+            var model = new FaviconModel();
+
+            //try loading a store specific favicon
+            var faviconFileName = string.Format("favicon-{0}.ico", _storeContext.CurrentStore.Id);
+            var localFaviconPath = System.IO.Path.Combine(_httpContext.Request.PhysicalApplicationPath, faviconFileName);
+            if (!System.IO.File.Exists(localFaviconPath))
+            {
+                //try loading a generic favicon
+                faviconFileName = "favicon.ico";
+                localFaviconPath = System.IO.Path.Combine(_httpContext.Request.PhysicalApplicationPath, faviconFileName);
+                if (!System.IO.File.Exists(localFaviconPath))
+                {
+                    return model;
+                }
+            }
+
+            model.FaviconUrl = _webHelper.GetStoreLocation() + faviconFileName;
+            return model;
+        }
+        
+        public virtual string PrepareRobotsTextFile()
+        {
+            var sb = new StringBuilder();
+
+            //if robots.custom.txt exists, let's use it instead of hard-coded data below
+            string robotsFilePath = System.IO.Path.Combine(CommonHelper.MapPath("~/"), "robots.custom.txt");
+            if (System.IO.File.Exists(robotsFilePath))
+            {
+                //the robots.txt file exists
+                string robotsFileContent = System.IO.File.ReadAllText(robotsFilePath);
+                sb.Append(robotsFileContent);
+            }
+            else
+            {
+                //doesn't exist. Let's generate it (default behavior)
+
+                var disallowPaths = new List<string>
+                {
+                    "/admin",
+                    "/bin/",
+                    "/content/files/",
+                    "/content/files/exportimport/",
+                    "/country/getstatesbycountryid",
+                    "/install",
+                    "/setproductreviewhelpfulness",
+                };
+                var localizableDisallowPaths = new List<string>
+                {
+                    "/addproducttocart/catalog/",
+                    "/addproducttocart/details/",
+                    "/backinstocksubscriptions/manage",
+                    "/boards/forumsubscriptions",
+                    "/boards/forumwatch",
+                    "/boards/postedit",
+                    "/boards/postdelete",
+                    "/boards/postcreate",
+                    "/boards/topicedit",
+                    "/boards/topicdelete",
+                    "/boards/topiccreate",
+                    "/boards/topicmove",
+                    "/boards/topicwatch",
+                    "/cart",
+                    "/checkout",
+                    "/checkout/billingaddress",
+                    "/checkout/completed",
+                    "/checkout/confirm",
+                    "/checkout/shippingaddress",
+                    "/checkout/shippingmethod",
+                    "/checkout/paymentinfo",
+                    "/checkout/paymentmethod",
+                    "/clearcomparelist",
+                    "/compareproducts",
+                    "/compareproducts/add/*",
+                    "/customer/avatar",
+                    "/customer/activation",
+                    "/customer/addresses",
+                    "/customer/changepassword",
+                    "/customer/checkusernameavailability",
+                    "/customer/downloadableproducts",
+                    "/customer/info",
+                    "/deletepm",
+                    "/emailwishlist",
+                    "/inboxupdate",
+                    "/newsletter/subscriptionactivation",
+                    "/onepagecheckout",
+                    "/order/history",
+                    "/orderdetails",
+                    "/passwordrecovery/confirm",
+                    "/poll/vote",
+                    "/privatemessages",
+                    "/returnrequest",
+                    "/returnrequest/history",
+                    "/rewardpoints/history",
+                    "/sendpm",
+                    "/sentupdate",
+                    "/shoppingcart/*",
+                    "/storeclosed",
+                    "/subscribenewsletter",
+                    "/topic/authenticate",
+                    "/viewpm",
+                    "/uploadfileproductattribute",
+                    "/uploadfilecheckoutattribute",
+                    "/wishlist",
+                };
+
+
+                const string newLine = "\r\n"; //Environment.NewLine
+                sb.Append("User-agent: *");
+                sb.Append(newLine);
+                //sitemaps
+                if (_localizationSettings.SeoFriendlyUrlsForLanguagesEnabled)
+                {
+                    //URLs are localizable. Append SEO code
+                    foreach (var language in _languageService.GetAllLanguages(storeId: _storeContext.CurrentStore.Id))
+                    {
+                        sb.AppendFormat("Sitemap: {0}{1}/sitemap.xml", _storeContext.CurrentStore.Url, language.UniqueSeoCode);
+                        sb.Append(newLine);
+                    }
+                }
+                else
+                {
+                    //localizable paths (without SEO code)
+                    sb.AppendFormat("Sitemap: {0}sitemap.xml", _storeContext.CurrentStore.Url);
+                    sb.Append(newLine);
+                }
+
+                //usual paths
+                foreach (var path in disallowPaths)
+                {
+                    sb.AppendFormat("Disallow: {0}", path);
+                    sb.Append(newLine);
+                }
+                //localizable paths (without SEO code)
+                foreach (var path in localizableDisallowPaths)
+                {
+                    sb.AppendFormat("Disallow: {0}", path);
+                    sb.Append(newLine);
+                }
+                if (_localizationSettings.SeoFriendlyUrlsForLanguagesEnabled)
+                {
+                    //URLs are localizable. Append SEO code
+                    foreach (var language in _languageService.GetAllLanguages(storeId: _storeContext.CurrentStore.Id))
+                    {
+                        foreach (var path in localizableDisallowPaths)
+                        {
+                            sb.AppendFormat("Disallow: /{0}{1}", language.UniqueSeoCode, path);
+                            sb.Append(newLine);
+                        }
+                    }
+                }
+
+                //load and add robots.txt additions to the end of file.
+                string robotsAdditionsFile = System.IO.Path.Combine(CommonHelper.MapPath("~/"), "robots.additions.txt");
+                if (System.IO.File.Exists(robotsAdditionsFile))
+                {
+                    string robotsFileContent = System.IO.File.ReadAllText(robotsAdditionsFile);
+                    sb.Append(robotsFileContent);
+                }
+            }
+
+            return sb.ToString();
+        }
+        
+        #endregion
+    }
+}
